# backend/openapi.yaml
openapi: 3.0.3
info:
  title: progctf — The Voyage API
  version: 0.1.0
  description: |
    Pirate-themed CTF backend (Bosun). Player REST API + admin API.
    JWT auth via httpOnly cookie (`progctf_session` for player, `progctf_admin` for admin)
    or `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:4000
    description: Local dev
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: progctf_session
    adminCookie:
      type: apiKey
      in: cookie
      name: progctf_admin
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, code, message]
      properties:
        ok: { type: boolean, example: false }
        code:
          type: string
          enum:
            - ERR_BAD_INPUT
            - ERR_BAD_FORMAT
            - ERR_BAD_CREDENTIALS
            - ERR_PIRATE_EXISTS
            - ERR_BANNED
            - ERR_NO_AUTH
            - ERR_BAD_AUTH
            - ERR_NOT_ADMIN
            - ERR_NO_REFRESH
            - ERR_BAD_REFRESH
            - ERR_NO_CREW
            - ERR_ALREADY_ON_CREW
            - ERR_CREW_NAME_TAKEN
            - ERR_CREW_BANNED
            - ERR_NO_ISLAND
            - ERR_NOT_PUBLISHED
            - ERR_NO_WHISPER
            - ERR_BAD_WHISPER
            - ERR_RATE_LIMITED
            - ERR_ALREADY_PLUNDERED
            - ERR_WRONG_TREASURE
            - ERR_VOYAGE_FROZEN
            - ERR_SLUG_TAKEN
            - ERR_WHISPER_EXISTS
            - ERR_NO_PIRATE
            - ERR_NO_UPDATES
            - ERR_SANDBOX_DOWN
        message: { type: string, example: "That ain't the right Treasure, sailor." }
        retryAfter: { type: integer }
...
    IslandList:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [port, open_sea, cursed_depths] }
        base_points: { type: integer }
        current_points: { type: integer }
        first_blood_crew: { type: string, nullable: true }
        solves: { type: integer }
        solved: { type: boolean }
    IslandDetail:
      allOf:
        - $ref: '#/components/schemas/IslandList'
        - type: object
          properties:
            description_md: { type: string }
            files_url: { type: string, nullable: true }
            sandbox_image: { type: string, nullable: true }
            solved_by_crew: { type: boolean }
            whispers:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  ordinal: { type: integer }
                  cost_points: { type: integer }
                  revealed: { type: boolean }
                  body_md: { type: string, nullable: true }
...
paths:
...
  /api/islands/:
    get:
      tags: [islands]
      summary: |
        List Islands. By default returns only `status='published'` (sanitized — no flag_hash).
        Admins may pass `?include_unpublished=1` to also see drafts and archived. Non-admins
        passing the param get 403 (we do NOT silently ignore — frontend needs to know).
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: query
          name: include_unpublished
          required: false
          description: |
            "1" or "true" — admin-only. Includes drafts and archived. 403 for non-admins.
          schema:
            type: string
            enum: ['1', 'true']
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  islands:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/IslandList'
                        - type: object
                          properties:
                            status: { type: string, enum: [draft, published, archived] }
        '403':
          description: non-admin tried to use include_unpublished
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/islands/{slug}:
    get:
      tags: [islands]
      summary: Island detail
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  island: { $ref: '#/components/schemas/IslandDetail' }
        '404': { description: not found / not published }
...
  /admin/islands/:
    post:
      tags: [admin]
      summary: Create Island
      security: [{ adminCookie: [] }, { bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, title, category, difficulty, description_md, base_points, flag]
              properties:
                slug: { type: string }
                title: { type: string }
                category: { type: string, enum: [cursed_ports, cipher_cove, shipwrights_forge, lighthouse, crows_nest, hidden_cargo, keymaster] }
                difficulty: { type: string, enum: [port, open_sea, cursed_depths] }
                description_md: { type: string }
                base_points: { type: integer, minimum: 50, maximum: 900 }
                flag: { type: string, pattern: '^progctf\{[a-z0-9_]+\}$' }
                files_url: { type: string, nullable: true }
                sandbox_image: { type: string, nullable: true }
                status: { type: string, enum: [draft, published, archived], default: draft }
      responses:
        '201': { description: created }
        '409': { description: slug taken }
  /admin/islands/{id}:
    patch:
      tags: [admin]
      summary: Edit Island (any subset of fields)
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: updated }
        '404': { description: not found }
    delete:
      tags: [admin]
      summary: Delete Island
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: scuttled }
  /admin/islands/{id}/files:
    post:
      tags: [admin]
      summary: Set the object-storage prefix for challenge files
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files_url]
              properties:
                files_url: { type: string, format: uri }
      responses:
        '200': { description: ok }
  /admin/islands/{id}/status:
    patch:
      tags: [admin]
      summary: Set Island status (draft / published / archived)
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [draft, published, archived] }
      responses:
        '200': { description: ok }
  /admin/islands/{id}/whispers:
    post:
      tags: [admin]
      summary: Add a Whisper
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ordinal, body_md, cost_points]
              properties:
                ordinal: { type: integer, minimum: 1, maximum: 3 }
                body_md: { type: string }
                cost_points: { type: integer, minimum: 0 }
      responses:
        '201': { description: created }
        '409': { description: whisper exists at that ordinal }
  /admin/islands/{id}/stats:
    get:
      tags: [admin]
      summary: Island stats — solves, attempts, first blood
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
...
  /admin/crews/{id}/ban:
    post:
      tags: [admin]
      summary: Ban a Crew
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
  /admin/crews/{id}/unban:
    post:
      tags: [admin]
      summary: Un-ban a Crew
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
  /admin/pirates/{id}/ban:
    post:
      tags: [admin]
      summary: Ban a Pirate
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
  /admin/submissions:
    get:
      tags: [admin]
      summary: Submission log (filterable)
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: query
          name: island_id
          schema: { type: string, format: uuid }
        - in: query
          name: crew_id
          schema: { type: string, format: uuid }
        - in: query
          name: pirate_id
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: ok }
  /admin/audit:
    get:
      tags: [admin]
      summary: Recent admin audit entries
      security: [{ adminCookie: [] }, { bearer: [] }]
      responses:
        '200': { description: ok }
...